+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All Notable changes to `event-store` will be documented in this file
 
+# 1.2
+
+* Added support for upcasting to multiple events
+* Deprecated support of upcasters returning non array values
+
 # 1.1.1
 
 Fixed partial stream upcasting